backend/ai_application/requirements.txt
langchain
langchain-community
langchain-ollama
fastapi
uvicorn
requests
python-dotenv
chromadb-client
pandas
sentence_transformers
